pyproject.toml
[tool.poetry]
name = "gam_bot"
version = "0.0.0"
description = "A fun discord bot for my discord server"
authors = ["Cameron Graybill <camerongraybill@gmail.com>", "Daron Arakelian <daronarakelian@gmail.com>"]
license = "MIT"
packages = [
    { include = "manage.py", from = "src" },
    { include = "gam", from = "src" },
    { include = "discord_bot", from = "src" },
    { include = "easy_messages", from = "src" },
    { include = "gam_coins", from = "src" },
    { include = "lmgtfy", from = "src" },
    { include = "social_score", from = "src" },
    { include = "dev_utils", from = "src" },
    { include = "advent_of_code", from = "src" },
]

[tool.poetry.scripts]
manage = "manage:main"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
discord-py = "^2.4.0"
yarl = "^1.11.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
django-stubs = "^5.0.4"
ruff = "^0.6.5"
psycopg = { extras = ["binary"], version = "^3.2.1" }


[tool.poetry.group.release]
optional = true

[tool.poetry.group.release.dependencies]
psycopg = { extras = ["c"], version = "^3.2.1" }
uvloop = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[tool.mypy]
files = "src"
plugins = ["mypy_django_plugin.main"]
strict = true
enable_incomplete_feature = ["NewGenericSyntax"]

[[tool.mypy.overrides]]
module = "uvloop.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "gam.settings.testing"

[tool.ruff]
target-version = "py312"
